# docs/swagger.yaml
openapi: 3.0.3
info:
  title: Mangaty API
  description: |
    API para la plataforma de cómics Mangaty.
    
    ## Características:
    - Autenticación segura con JWT
    - Sistema de roles flexible (lector → autor → admin)
    - Microeconomía con TyCoins
    - Subida de imágenes con Cloudinary
    - Sistema de recomendaciones con IA
    
    ## Roles del sistema:
    - **lector**: Rol por defecto. Lee cómics y desbloquea capítulos
    - **autor**: Puede crear cómics y publicar capítulos. Cualquier lector puede subir contenido
    - **administrador**: Control total. Asigna roles y elimina contenido
    
    ## Flujo de autenticación:
    1. Registrar usuario → `POST /api/v1/auth/register`
    2. Iniciar sesión → `POST /api/v1/auth/login`
    3. Clic en **Authorize** y pegar: `Bearer <token>`
  version: 1.0.0
  contact:
    name: Mangaty Team
    email: support@mangaty.com
servers:
  - url: http://localhost:8080
    description: Desarrollo local
  - url: https://api.mangaty.com
    description: Producción
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token JWT del login. Formato: `Bearer <token>`
        El token incluye los roles del usuario
  schemas:
...
    StandardResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
        timestamp:
          type: string
          format: date-time

    # ============ AUTH ============
...
    UserProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
          format: email
        roles:
          type: array
          items:
            type: string
          example: ["lector", "autor"]
        created_at:
          type: string
          format: date-time
...
    AssignRoleRequest:
      type: object
      required: [user_id, role]
      properties:
        user_id:
          type: string
          format: uuid
          description: ID del usuario al que asignar el rol
        role:
          type: string
          enum: [autor, administrador]
          description: Rol a asignar

    # ============ WALLET ============
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
        timestamp:
          type: string
          format: date-time
  responses:
...
    Forbidden:
      description: Sin permisos
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
paths:
...
  /api/v1/user/profile:
    get:
      tags: [User]
      summary: Ver perfil
      description: Obtiene perfil con roles actuales
      operationId: getProfile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Perfil del usuario
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/StandardResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UserProfile'


  # ============ ADMIN ============
  /api/v1/admin/assign-role:
    post:
      tags: [Admin]
      summary: Asignar rol a usuario
      description: Solo **administradores**. Asigna rol autor o administrador a otro usuario
      operationId: assignRole
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignRoleRequest'
            example:
              user_id: "550e8400-e29b-41d4-a716-446655440000"
              role: "autor"
      responses:
        '200':
          description: Rol asignado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardResponse'
        '403':
          $ref: '#/components/responses/Forbidden'

  # ============ WALLET ============
...
security: []